move ListFieldTests to right package
