+++ src/test/java/com/celements/model/classes/fields/list/AccessRightLevelsFieldTest.java
@@ -1,4 +1,4 @@
-package com.celements.model.classes.fields;
+package com.celements.model.classes.fields.list;
 
 import static com.celements.common.test.CelementsTestUtils.*;
 import static org.junit.Assert.*;
@@ -16,28 +16,24 @@ import org.xwiki.model.reference.DocumentReference;
 import com.celements.common.test.AbstractComponentTest;
 import com.celements.model.access.IModelAccessFacade;
 import com.celements.model.classes.TestClassDefinition;
+import com.celements.model.classes.fields.list.AccessRightLevelsField;
 import com.celements.model.classes.fields.list.EnumListField;
 import com.celements.model.util.ClassFieldValue;
+import com.celements.rights.access.EAccessLevel;
+import com.google.common.base.Joiner;
 import com.xpn.xwiki.doc.XWikiDocument;
 import com.xpn.xwiki.objects.classes.BaseClass;
+import com.xpn.xwiki.objects.classes.LevelsClass;
 import com.xpn.xwiki.objects.classes.PropertyClass;
-import com.xpn.xwiki.objects.classes.StaticListClass;
 import com.xpn.xwiki.web.Utils;
 
-public class EnumListFieldTest extends AbstractComponentTest {
+public class AccessRightLevelsFieldTest extends AbstractComponentTest {
 
-  private enum TestEnum {
-    A,
-    B,
-    C,
-    D;
-  }
-
-  private EnumListField.Builder<TestEnum> fieldBuilder;
+  private AccessRightLevelsField.Builder fieldBuilder;
 
   @Before
   public void prepareTest() throws Exception {
-    fieldBuilder = new EnumListField.Builder<>(TestClassDefinition.NAME, "name", TestEnum.class);
+    fieldBuilder = new AccessRightLevelsField.Builder(TestClassDefinition.NAME, "name");
   }
 
   @Test
@@ -47,57 +43,59 @@ public class EnumListFieldTest extends AbstractComponentTest {
 
   @Test
   public void test_getXField() throws Exception {
-    EnumListField<TestEnum> field = fieldBuilder.build();
-    assertTrue(field.getXField() instanceof StaticListClass);
-    StaticListClass xField = (StaticListClass) field.getXField();
-    assertEquals("A|B|C|D", xField.getValues());
+    AccessRightLevelsField field = fieldBuilder.build();
+    System.out.println(field.getXField());
+    assertTrue(field.getXField() instanceof LevelsClass);
   }
 
   @Test
   public void test_resolve_serialize() throws Exception {
-    EnumListField<TestEnum> field = fieldBuilder.build();
+    AccessRightLevelsField field = fieldBuilder.build();
     DocumentReference classRef = field.getClassDef().getClassRef();
     IModelAccessFacade modelAccess = Utils.getComponent(IModelAccessFacade.class);
     XWikiDocument doc = new XWikiDocument(classRef);
-    TestEnum value = TestEnum.B;
+    EAccessLevel value = EAccessLevel.VIEW;
 
     BaseClass bClass = expectNewBaseObject(classRef);
     expectPropertyClass(bClass, field.getName(), (PropertyClass) field.getXField());
 
     replayDefault();
     modelAccess.setProperty(doc, new ClassFieldValue<>(field, Arrays.asList(value)));
-    List<TestEnum> ret = modelAccess.getProperty(doc, field);
+    List<EAccessLevel> ret = modelAccess.getProperty(doc, field);
     verifyDefault();
 
     assertEquals(Arrays.asList(value), ret);
-    assertEquals(value.name(), modelAccess.getXObject(doc, classRef).getStringValue(
+    assertEquals(value.getIdentifier(), modelAccess.getXObject(doc, classRef).getStringValue(
         field.getName()));
   }
 
   @Test
   public void test_resolve_serialize_multiselect() throws Exception {
-    EnumListField<TestEnum> field = fieldBuilder.multiSelect(true).build();
+    AccessRightLevelsField field = ((AccessRightLevelsField.Builder) fieldBuilder.multiSelect(
+        true)).build();
     DocumentReference classRef = field.getClassDef().getClassRef();
     IModelAccessFacade modelAccess = Utils.getComponent(IModelAccessFacade.class);
     XWikiDocument doc = new XWikiDocument(classRef);
-    List<TestEnum> value = Arrays.asList(TestEnum.B, TestEnum.D);
+    List<EAccessLevel> value = Arrays.asList(EAccessLevel.VIEW, EAccessLevel.EDIT);
 
     BaseClass bClass = expectNewBaseObject(classRef);
     expectPropertyClass(bClass, field.getName(), (PropertyClass) field.getXField());
 
     replayDefault();
     modelAccess.setProperty(doc, new ClassFieldValue<>(field, value));
-    List<TestEnum> ret = modelAccess.getProperty(doc, field);
+    List<EAccessLevel> ret = modelAccess.getProperty(doc, field);
     verifyDefault();
 
     assertEquals(value, ret);
-    assertEquals(Arrays.asList("B", "D"), modelAccess.getXObject(doc, classRef).getListValue(
-        field.getName()));
+    assertEquals(Joiner.on(',').join(EAccessLevel.VIEW.getIdentifier(),
+        EAccessLevel.EDIT.getIdentifier()), modelAccess.getXObject(doc, classRef).getStringValue(
+            field.getName()));
   }
 
   @Test
   public void test_resolve_serialize_null() throws Exception {
-    EnumListField<TestEnum> field = fieldBuilder.multiSelect(true).build();
+    AccessRightLevelsField field = ((AccessRightLevelsField.Builder) fieldBuilder.multiSelect(
+        true)).build();
     DocumentReference classRef = field.getClassDef().getClassRef();
     IModelAccessFacade modelAccess = Utils.getComponent(IModelAccessFacade.class);
     XWikiDocument doc = new XWikiDocument(classRef);
@@ -106,16 +104,16 @@ public class EnumListFieldTest extends AbstractComponentTest {
     expectPropertyClass(bClass, field.getName(), (PropertyClass) field.getXField());
 
     replayDefault();
-    List<TestEnum> ret1 = modelAccess.getProperty(doc, field);
+    List<EAccessLevel> ret1 = modelAccess.getProperty(doc, field);
     modelAccess.setProperty(doc, new ClassFieldValue<>(field, null));
-    List<TestEnum> ret2 = modelAccess.getProperty(doc, field);
+    List<EAccessLevel> ret2 = modelAccess.getProperty(doc, field);
     verifyDefault();
 
     assertNotNull(ret1);
     assertTrue(ret1.isEmpty());
     assertNotNull(ret2);
     assertTrue(ret2.isEmpty());
-    assertTrue(modelAccess.getXObject(doc, classRef).getListValue(field.getName()).isEmpty());
+    assertTrue(modelAccess.getXObject(doc, classRef).getStringValue(field.getName()).isEmpty());
   }
 
 }
