+++ src/test/java/com/celements/model/classes/fields/list/EnumListFieldTest.java
@@ -1,4 +1,4 @@
-package com.celements.model.classes.fields;
+package com.celements.model.classes.fields.list;
 
 import static com.celements.common.test.CelementsTestUtils.*;
 import static org.junit.Assert.*;
@@ -16,24 +16,28 @@ import org.xwiki.model.reference.DocumentReference;
 import com.celements.common.test.AbstractComponentTest;
 import com.celements.model.access.IModelAccessFacade;
 import com.celements.model.classes.TestClassDefinition;
-import com.celements.model.classes.fields.list.AccessRightLevelsField;
 import com.celements.model.classes.fields.list.EnumListField;
 import com.celements.model.util.ClassFieldValue;
-import com.celements.rights.access.EAccessLevel;
-import com.google.common.base.Joiner;
 import com.xpn.xwiki.doc.XWikiDocument;
 import com.xpn.xwiki.objects.classes.BaseClass;
-import com.xpn.xwiki.objects.classes.LevelsClass;
 import com.xpn.xwiki.objects.classes.PropertyClass;
+import com.xpn.xwiki.objects.classes.StaticListClass;
 import com.xpn.xwiki.web.Utils;
 
-public class AccessRightLevelsFieldTest extends AbstractComponentTest {
+public class EnumListFieldTest extends AbstractComponentTest {
 
-  private AccessRightLevelsField.Builder fieldBuilder;
+  private enum TestEnum {
+    A,
+    B,
+    C,
+    D;
+  }
+
+  private EnumListField.Builder<TestEnum> fieldBuilder;
 
   @Before
   public void prepareTest() throws Exception {
-    fieldBuilder = new AccessRightLevelsField.Builder(TestClassDefinition.NAME, "name");
+    fieldBuilder = new EnumListField.Builder<>(TestClassDefinition.NAME, "name", TestEnum.class);
   }
 
   @Test
@@ -43,59 +47,57 @@ public class AccessRightLevelsFieldTest extends AbstractComponentTest {
 
   @Test
   public void test_getXField() throws Exception {
-    AccessRightLevelsField field = fieldBuilder.build();
-    System.out.println(field.getXField());
-    assertTrue(field.getXField() instanceof LevelsClass);
+    EnumListField<TestEnum> field = fieldBuilder.build();
+    assertTrue(field.getXField() instanceof StaticListClass);
+    StaticListClass xField = (StaticListClass) field.getXField();
+    assertEquals("A|B|C|D", xField.getValues());
   }
 
   @Test
   public void test_resolve_serialize() throws Exception {
-    AccessRightLevelsField field = fieldBuilder.build();
+    EnumListField<TestEnum> field = fieldBuilder.build();
     DocumentReference classRef = field.getClassDef().getClassRef();
     IModelAccessFacade modelAccess = Utils.getComponent(IModelAccessFacade.class);
     XWikiDocument doc = new XWikiDocument(classRef);
-    EAccessLevel value = EAccessLevel.VIEW;
+    TestEnum value = TestEnum.B;
 
     BaseClass bClass = expectNewBaseObject(classRef);
     expectPropertyClass(bClass, field.getName(), (PropertyClass) field.getXField());
 
     replayDefault();
     modelAccess.setProperty(doc, new ClassFieldValue<>(field, Arrays.asList(value)));
-    List<EAccessLevel> ret = modelAccess.getProperty(doc, field);
+    List<TestEnum> ret = modelAccess.getFieldValue(doc, field).orNull();
     verifyDefault();
 
     assertEquals(Arrays.asList(value), ret);
-    assertEquals(value.getIdentifier(), modelAccess.getXObject(doc, classRef).getStringValue(
+    assertEquals(value.name(), modelAccess.getXObject(doc, classRef).getStringValue(
         field.getName()));
   }
 
   @Test
   public void test_resolve_serialize_multiselect() throws Exception {
-    AccessRightLevelsField field = ((AccessRightLevelsField.Builder) fieldBuilder.multiSelect(
-        true)).build();
+    EnumListField<TestEnum> field = fieldBuilder.multiSelect(true).build();
     DocumentReference classRef = field.getClassDef().getClassRef();
     IModelAccessFacade modelAccess = Utils.getComponent(IModelAccessFacade.class);
     XWikiDocument doc = new XWikiDocument(classRef);
-    List<EAccessLevel> value = Arrays.asList(EAccessLevel.VIEW, EAccessLevel.EDIT);
+    List<TestEnum> value = Arrays.asList(TestEnum.B, TestEnum.D);
 
     BaseClass bClass = expectNewBaseObject(classRef);
     expectPropertyClass(bClass, field.getName(), (PropertyClass) field.getXField());
 
     replayDefault();
     modelAccess.setProperty(doc, new ClassFieldValue<>(field, value));
-    List<EAccessLevel> ret = modelAccess.getProperty(doc, field);
+    List<TestEnum> ret = modelAccess.getFieldValue(doc, field).orNull();
     verifyDefault();
 
     assertEquals(value, ret);
-    assertEquals(Joiner.on(',').join(EAccessLevel.VIEW.getIdentifier(),
-        EAccessLevel.EDIT.getIdentifier()), modelAccess.getXObject(doc, classRef).getStringValue(
-            field.getName()));
+    assertEquals(Arrays.asList("B", "D"), modelAccess.getXObject(doc, classRef).getListValue(
+        field.getName()));
   }
 
   @Test
   public void test_resolve_serialize_null() throws Exception {
-    AccessRightLevelsField field = ((AccessRightLevelsField.Builder) fieldBuilder.multiSelect(
-        true)).build();
+    EnumListField<TestEnum> field = fieldBuilder.multiSelect(true).build();
     DocumentReference classRef = field.getClassDef().getClassRef();
     IModelAccessFacade modelAccess = Utils.getComponent(IModelAccessFacade.class);
     XWikiDocument doc = new XWikiDocument(classRef);
@@ -104,16 +106,16 @@ public class AccessRightLevelsFieldTest extends AbstractComponentTest {
     expectPropertyClass(bClass, field.getName(), (PropertyClass) field.getXField());
 
     replayDefault();
-    List<EAccessLevel> ret1 = modelAccess.getProperty(doc, field);
+    List<TestEnum> ret1 = modelAccess.getFieldValue(doc, field).orNull();
     modelAccess.setProperty(doc, new ClassFieldValue<>(field, null));
-    List<EAccessLevel> ret2 = modelAccess.getProperty(doc, field);
+    List<TestEnum> ret2 = modelAccess.getFieldValue(doc, field).orNull();
     verifyDefault();
 
     assertNotNull(ret1);
     assertTrue(ret1.isEmpty());
     assertNotNull(ret2);
     assertTrue(ret2.isEmpty());
-    assertTrue(modelAccess.getXObject(doc, classRef).getStringValue(field.getName()).isEmpty());
+    assertTrue(modelAccess.getXObject(doc, classRef).getListValue(field.getName()).isEmpty());
   }
 
 }
